Changes to settings processing on backend
I added some more settings, so I updated the backend to reflect these
changes. It should now pass them appropriately to the front end (which
as of now stores them and then does nothing with them.)

It should be noted that the way the settings are embeded in the page is
not very scalable, but it should work for any finite number of settings.